[sinks] Move ComputeSink structs out of storage
