+++ src/compute-client/build.rs
@@ -46,6 +46,7 @@ fn main() {
                 "compute-client/src/plan/top_k.proto",
                 "compute-client/src/response.proto",
                 "compute-client/src/service.proto",
+                "compute-client/src/sinks.proto",
             ],
             &[".."],
         )
