+++ src/compute-client/src/command.proto
@@ -11,6 +11,7 @@ syntax = "proto3";
 
 import "compute-client/src/logging.proto";
 import "compute-client/src/plan.proto";
+import "compute-client/src/sinks.proto";
 import "expr/src/linear.proto";
 import "expr/src/relation.proto";
 import "expr/src/scalar.proto";
@@ -19,9 +20,8 @@ import "repr/src/antichain.proto";
 import "repr/src/global_id.proto";
 import "repr/src/relation_and_scalar.proto";
 import "repr/src/row.proto";
-import "storage/src/protocol/client.proto";
 import "storage/src/controller.proto";
-import "storage/src/types/sinks.proto";
+import "storage/src/protocol/client.proto";
 
 import "google/protobuf/empty.proto";
 
@@ -74,7 +74,7 @@ message ProtoDataflowDescription {
 
     message ProtoSinkExport {
         mz_repr.global_id.ProtoGlobalId id = 1;
-        mz_storage.types.sinks.ProtoComputeSinkDesc sink_desc = 2;
+        mz_compute_client.sinks.ProtoComputeSinkDesc sink_desc = 2;
     }
 
     repeated ProtoSourceImport source_imports = 1;
