+++ src/compute-client/src/lib.rs
@@ -20,3 +20,4 @@ pub mod logging;
 pub mod plan;
 pub mod response;
 pub mod service;
+pub mod sinks;
