+++ src/compute-client/src/sinks.proto
@@ -1,0 +1,42 @@
+// Copyright Materialize, Inc. and contributors. All rights reserved.
+//
+// Use of this software is governed by the Business Source License
+// included in the LICENSE file.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0.
+syntax = "proto3";
+
+import "google/protobuf/empty.proto";
+
+import "repr/src/antichain.proto";
+import "repr/src/global_id.proto";
+import "repr/src/relation_and_scalar.proto";
+import "storage/src/controller.proto";
+
+package mz_compute_client.sinks;
+
+message ProtoComputeSinkDesc {
+    mz_repr.global_id.ProtoGlobalId from = 1;
+    mz_repr.relation_and_scalar.ProtoRelationDesc from_desc = 2;
+    ProtoComputeSinkConnection connection = 3;
+    ProtoSinkAsOf as_of = 4;
+}
+
+message ProtoComputeSinkConnection {
+    oneof kind {
+        google.protobuf.Empty tail = 1;
+        ProtoPersistSinkConnection persist = 2;
+    }
+}
+
+message ProtoSinkAsOf {
+    mz_repr.antichain.ProtoU64Antichain frontier = 1;
+    bool strict = 2;
+}
+
+message ProtoPersistSinkConnection {
+    mz_repr.relation_and_scalar.ProtoRelationDesc value_desc = 1;
+    mz_storage.controller.ProtoCollectionMetadata storage_metadata = 2;
+}
