+++ src/compute-client/src/sinks.rs
@@ -1,0 +1,184 @@
+// Copyright Materialize, Inc. and contributors. All rights reserved.
+//
+// Use of this software is governed by the Business Source License
+// included in the LICENSE file.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0.
+
+use proptest::prelude::{any, Arbitrary, BoxedStrategy, Strategy};
+use proptest_derive::Arbitrary;
+use serde::{Deserialize, Serialize};
+use timely::progress::Antichain;
+
+use mz_proto::{IntoRustIfSome, ProtoType, RustType, TryFromProtoError};
+use mz_repr::{GlobalId, RelationDesc};
+use mz_storage::controller::CollectionMetadata;
+
+include!(concat!(env!("OUT_DIR"), "/mz_compute_client.sinks.rs"));
+
+/// A sink for updates to a relational collection.
+#[derive(Clone, Debug, Serialize, Deserialize, Eq, PartialEq)]
+pub struct ComputeSinkDesc<S = (), T = mz_repr::Timestamp> {
+    pub from: GlobalId,
+    pub from_desc: RelationDesc,
+    pub connection: ComputeSinkConnection<S>,
+    pub as_of: SinkAsOf<T>,
+}
+
+impl Arbitrary for ComputeSinkDesc<CollectionMetadata, mz_repr::Timestamp> {
+    type Strategy = BoxedStrategy<Self>;
+    type Parameters = ();
+
+    fn arbitrary_with(_: Self::Parameters) -> Self::Strategy {
+        (
+            any::<GlobalId>(),
+            any::<RelationDesc>(),
+            any::<ComputeSinkConnection<CollectionMetadata>>(),
+            any::<SinkAsOf<mz_repr::Timestamp>>(),
+        )
+            .prop_map(|(from, from_desc, connection, as_of)| ComputeSinkDesc {
+                from,
+                from_desc,
+                connection,
+                as_of,
+            })
+            .boxed()
+    }
+}
+
+impl RustType<ProtoComputeSinkDesc> for ComputeSinkDesc<CollectionMetadata, mz_repr::Timestamp> {
+    fn into_proto(&self) -> ProtoComputeSinkDesc {
+        ProtoComputeSinkDesc {
+            connection: Some(self.connection.into_proto()),
+            from: Some(self.from.into_proto()),
+            from_desc: Some(self.from_desc.into_proto()),
+            as_of: Some(self.as_of.into_proto()),
+        }
+    }
+
+    fn from_proto(proto: ProtoComputeSinkDesc) -> Result<Self, TryFromProtoError> {
+        Ok(ComputeSinkDesc {
+            from: proto.from.into_rust_if_some("ProtoComputeSinkDesc::from")?,
+            from_desc: proto
+                .from_desc
+                .into_rust_if_some("ProtoComputeSinkDesc::from_desc")?,
+            connection: proto
+                .connection
+                .into_rust_if_some("ProtoComputeSinkDesc::connection")?,
+            as_of: proto
+                .as_of
+                .into_rust_if_some("ProtoComputeSinkDesc::as_of")?,
+        })
+    }
+}
+
+#[derive(Arbitrary, Clone, Debug, Serialize, Deserialize, Eq, PartialEq)]
+pub enum ComputeSinkConnection<S = ()> {
+    Tail(TailSinkConnection),
+    Persist(PersistSinkConnection<S>),
+}
+
+impl<S> ComputeSinkConnection<S> {
+    /// Returns the name of the sink connection.
+    pub fn name(&self) -> &'static str {
+        match self {
+            ComputeSinkConnection::Tail(_) => "tail",
+            ComputeSinkConnection::Persist(_) => "persist",
+        }
+    }
+}
+
+impl RustType<ProtoComputeSinkConnection> for ComputeSinkConnection<CollectionMetadata> {
+    fn into_proto(&self) -> ProtoComputeSinkConnection {
+        use proto_compute_sink_connection::Kind;
+        ProtoComputeSinkConnection {
+            kind: Some(match self {
+                ComputeSinkConnection::Tail(_tail) => Kind::Tail(()),
+                ComputeSinkConnection::Persist(persist) => Kind::Persist(persist.into_proto()),
+            }),
+        }
+    }
+
+    fn from_proto(proto: ProtoComputeSinkConnection) -> Result<Self, TryFromProtoError> {
+        use proto_compute_sink_connection::Kind;
+        let kind = proto
+            .kind
+            .ok_or_else(|| TryFromProtoError::missing_field("ProtoComputeSinkConnection::kind"))?;
+        Ok(match kind {
+            Kind::Tail(_tail) => ComputeSinkConnection::Tail(TailSinkConnection {}),
+            Kind::Persist(persist) => ComputeSinkConnection::Persist(persist.into_rust()?),
+        })
+    }
+}
+
+#[derive(Arbitrary, Default, Clone, Debug, Serialize, Deserialize, Eq, PartialEq)]
+pub struct TailSinkConnection {}
+
+#[derive(Arbitrary, Clone, Debug, Eq, PartialEq, Serialize, Deserialize)]
+pub struct PersistSinkConnection<S> {
+    pub value_desc: RelationDesc,
+    pub storage_metadata: S,
+}
+
+impl RustType<ProtoPersistSinkConnection> for PersistSinkConnection<CollectionMetadata> {
+    fn into_proto(&self) -> ProtoPersistSinkConnection {
+        ProtoPersistSinkConnection {
+            value_desc: Some(self.value_desc.into_proto()),
+            storage_metadata: Some(self.storage_metadata.into_proto()),
+        }
+    }
+
+    fn from_proto(proto: ProtoPersistSinkConnection) -> Result<Self, TryFromProtoError> {
+        Ok(PersistSinkConnection {
+            value_desc: proto
+                .value_desc
+                .into_rust_if_some("ProtoPersistSinkConnection::value_desc")?,
+            storage_metadata: proto
+                .storage_metadata
+                .into_rust_if_some("ProtoPersistSinkConnection::storage_metadata")?,
+        })
+    }
+}
+
+#[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize)]
+pub struct SinkAsOf<T = mz_repr::Timestamp> {
+    pub frontier: Antichain<T>,
+    pub strict: bool,
+}
+
+impl Arbitrary for SinkAsOf<mz_repr::Timestamp> {
+    type Strategy = BoxedStrategy<Self>;
+    type Parameters = ();
+
+    fn arbitrary_with(_: Self::Parameters) -> Self::Strategy {
+        (
+            proptest::collection::vec(any::<mz_repr::Timestamp>(), 1..4),
+            any::<bool>(),
+        )
+            .prop_map(|(frontier, strict)| SinkAsOf {
+                frontier: Antichain::from(frontier),
+                strict,
+            })
+            .boxed()
+    }
+}
+
+impl RustType<ProtoSinkAsOf> for SinkAsOf<mz_repr::Timestamp> {
+    fn into_proto(&self) -> ProtoSinkAsOf {
+        ProtoSinkAsOf {
+            frontier: Some(self.frontier.into_proto()),
+            strict: self.strict,
+        }
+    }
+
+    fn from_proto(proto: ProtoSinkAsOf) -> Result<Self, TryFromProtoError> {
+        Ok(SinkAsOf {
+            frontier: proto
+                .frontier
+                .into_rust_if_some("ProtoSinkAsOf::frontier")?,
+            strict: proto.strict,
+        })
+    }
+}
